added some new api, fixed some librart issues

+++ tervel/util/progress_assurance.cc
@@ -28,7 +28,7 @@ THE SOFTWARE.
 namespace tervel {
 namespace util {
 
-void ProgressAssurance::p_check_for_announcement(size_t &help_id) {
+void ProgressAssurance::p_check_for_announcement(int64_t &help_id) {
     help_id++;
     if (help_id >= num_threads_) {
       help_id = 0;
